# perf(compiler-cli): template type-checking performance improvements