See individual commits.

---

I tested these changes in the [ng-speed-rebuild repo](https://github.com/elvisbegovic/ng-speed-rebuild), using Angular 10.1-next.5 and the changes in this PR, using roughly the following:

```bash
nvm use 12.16.1
git clone https://github.com/elvisbegovic/ng-speed-rebuild.git
cd ng-speed-rebuild
yarn
yarn ng update @angular/core @angular/cli --next --force --allow-dirty
# Install changes from PR
yarn ng build # Triggers ngcc
# Edit projects/one/tsconfig.app.json to enable "ivyTemplateTypeCheck", "fullTemplateTypeCheck" and "strictTemplates"
time ./node_modules/.bin/ngc -p projects/one/tsconfig.app.json
```

> Note that the above configuration will actually cause compilation errors due to `strictTemplates` being enabled, which means that there's no code being emitted. The emit phase is expensive by itself (especially with source maps enabled) but not included in the below figures.

Prior to this PR, using 10.1-next.5, `ngc` took **75 seconds** to execute. With the changes from this PR, it is reduced to **45 seconds**; a **30 second** reduction in time. The 30 second improvement is mostly from reduced type-checking time that TypeScript performs; from 35s to 7s. This is beneficial for incremental builds as well, given that type-checking is not done incrementally and has to be repeated from scratch on each incremental compilation. Side-note: the CLI performs incremental type-checking in a background process, so the effect of this may not be immediately visible (although your CPU will enjoy the extra free time!).

---

Thanks to @elvisbegovic for the awesome ng-speed-rebuild repo ❤️ Having these repo's available is super valuable in chasing down performance bottlenecks and experimenting with ways to address them.